refactor a lot of messaging to accommodate misc crap needed -- still …

+++ client/src/contexts/AuthContext.js
@@ -1,5 +1,5 @@
 import React, { useContext, useState, useEffect } from 'react';
-import { auth } from '../firebase'
+import { auth, firestore } from '../firebase'
 
 const AuthContext = React.createContext()
 
@@ -10,6 +10,7 @@ export function useAuth() {
 export function AuthProvider({ children }) {
   const [currentUser, setCurrentUser] = useState()
   const [loading, setLoading] = useState(true)
+  const [userRole, setUserRole] = useState(null); // Add state for user role
 
   function signup(email, password) {
     return auth.createUserWithEmailAndPassword(email, password)
@@ -28,24 +29,44 @@ export function AuthProvider({ children }) {
   }
 
   useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged(user => {
-      setCurrentUser(user)
-      setLoading(false)
+
+    const unsubscribe = auth.onAuthStateChanged(async user => {
+      setCurrentUser(user);
+      setLoading(false);
 
       // gonna pop localStorageCurrentUserID into local storage to make it
       // accessible everywhere...
+      // same with localStorageCurrentUserRole, as that will be handy
       if (user) {
-        localStorage.setItem('localStorageCurrentUserID', user.uid);
+        try {
+          const userRef = firestore.collection('users').doc(user.uid);
+          const doc = await userRef.get();
+          if (doc.exists) {
+            const userData = doc.data();
+            setUserRole(userData.role); // Set the user role state
+            localStorage.setItem('localStorageCurrentUserID', user.uid);
+            localStorage.setItem('localStorageCurrentUserRole', userData.role);
+          } else {
+            console.log('No such document!');
+            setUserRole(null);
+          }
+        } catch (error) {
+          console.error('Error fetching user role:', error);
+        }
       } else {
         localStorage.removeItem('localStorageCurrentUserID');
+        localStorage.removeItem('localStorageCurrentUserRole');
+        setUserRole(null);
       }
+    });
 
-    })
-    return unsubscribe
-  }, [])
+    return unsubscribe;
+
+  }, []);
 
   const value = {
     currentUser,
+    userRole,
     signup,
     login,
     logout,
